Good enough to watch S2.
The editing was strange (cuts in weird places forward and backward in
time) and the pacing was off (the second to last episode was way too
frantic and could have been the finale, which it seems like matches
some others' opinions).

So the wrong person one I think, but it was enjoyable.